More info page modification update
Addition of modifying the maori word, description, and the category of the word. Also implemented the date added. If the word is already pre-existing on the database, it will show it as "not recorded" but if a word is made new, the date that is is added will be displayed